dev-libs/libcgroup/cgconfig.initd: Major refactorization
This is a complete refactorization using POSIX shell, instead of bash;
and use a simpler set up functions to initialize CGroups;
Offer the option to use PID classification on start up by setting
CG_CLASSIFY=Yes in the configuration file (which was mandatory in the
previous iterration of the init script.)

Some details...

The functionalities of move_all_to_init_class() is completely removed
because there is no need to do this complicated set up when inheriting
CGroups is available, and, is used in CGroup set up by package like
OpenRC. And this can be setted per binary basis instead of messing with
ALL the PIDs.

So, the helpers related to moving PIDS to the top CGroups hierarchy, like
parse_mount() are also removed.

And thus remain only a simple CGroup set up which use cgconfig.conf file
to automatically set up CGroups hierarchies (system/user groups.)
And then, manage unmounting user defined CGroups hierarchy.

This is explain why some many lines are removed and... remain only a
very handy functions.